add delegate list
