+++ app/src/main/java/com/khaymoev/myapplication/models/ArticleItem.kt
@@ -5,7 +5,7 @@ import java.util.*
 //      Заголовок, кнопка лайка, кол-во лайков, автор
 data class ArticleItem(
     val title: String,
-    val amountLike: Int,
+    var amountLike: Int,
     val author: String,
     val uuid: UUID,
     val email: String = ""
