+++ app/src/main/java/com/khaymoev/myapplication/ui/main_page/delegates/ArticleImageItemDelegate.kt
@@ -1,4 +1,4 @@
-package com.khaymoev.myapplication.ui.main_page
+package com.khaymoev.myapplication.ui.main_page.delegates
 
 import android.view.LayoutInflater
 import android.view.View
@@ -6,48 +6,46 @@ import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import com.hannesdorfmann.adapterdelegates4.AbsListItemAdapterDelegate
 import com.khaymoev.myapplication.R
-import com.khaymoev.myapplication.models.ArticleItem
+import com.khaymoev.myapplication.models.ArticleImageItem
 import kotlinx.android.extensions.LayoutContainer
-import kotlinx.android.synthetic.main.item_article.view.*
+import kotlinx.android.synthetic.main.item_article_image.view.*
 
 /**
  * @author Vasiliy Khaymoev on 27.09.2021.
  */
-class ArticleItemDelegate(
-    private val onItemClick: (item: ArticleItem) -> Unit
-) : AbsListItemAdapterDelegate<Any, Any, ArticleItemDelegate.ViewHolder>() {
+class ArticleImageItemDelegate(
+    private val onItemClick: (item: ArticleImageItem) -> Unit
+) : AbsListItemAdapterDelegate<Any, Any, ArticleImageItemDelegate.ViewHolder>() {
 
     override fun isForViewType(item: Any, items: MutableList<Any>, position: Int): Boolean {
-        return item is ArticleItem
+        return item is ArticleImageItem
     }
 
     override fun onCreateViewHolder(parent: ViewGroup): ViewHolder {
         val itemView = LayoutInflater.from(parent.context)
-            .inflate(R.layout.item_article, parent, false)
+            .inflate(R.layout.item_article_image, parent, false)
         return ViewHolder(itemView, onItemClick)
     }
 
     override fun onBindViewHolder(item: Any, viewHolder: ViewHolder, payloads: MutableList<Any>) {
-        viewHolder.bind(item as ArticleItem)
+        viewHolder.bind(item as ArticleImageItem)
     }
 
     inner class ViewHolder(
         override val containerView: View,
-        private val onItemClick: (item: ArticleItem) -> Unit
+        private val onItemClick: (item: ArticleImageItem) -> Unit
     ) : RecyclerView.ViewHolder(containerView), LayoutContainer {
 
-        private var currentItem: ArticleItem? = null
+        private var currentItem: ArticleImageItem? = null
 
         init {
             containerView.setOnClickListener { currentItem?.let(onItemClick) }
         }
 
-        fun bind(item: ArticleItem) = with(containerView) {
+        fun bind(item: ArticleImageItem) = with(containerView) {
             currentItem = item
-            title.text = item.title
-            amountLike.text = item.amountLike.toString()
-            author.text = item.author
-            email.text = item.email
+            title_image.text = item.title
+            author_image.text = item.author
         }
     }
 }
